requirements.txt
fastapi
uvicorn
pydantic
beautifulsoup4
requests
langchain
langchain-community
langchain-openai
faiss-cpu
sentence-transformers
python-dotenv
pytest
httpx
pandas
